Allow raw pixel scrolling in web contents in Aura.
A future patch will allow raw pixel scrolling in views.

ui::LocatedEvent stores a floating point location. Currently
integral data is returned for all previously existing
accessors. In the future, all accessors should return the
floating point values.

BUG=329354
TEST=GestureRecognizerTest.GestureEventScroll

Review URL: https://codereview.chromium.org/143603002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@247358 0039d316-1c4b-4281-b951-d872f2087c98

+++ ui/events/gestures/velocity_calculator.h
@@ -18,7 +18,7 @@ class EVENTS_EXPORT VelocityCalculator {
  public:
   explicit VelocityCalculator(int bufferSize);
   ~VelocityCalculator();
-  void PointSeen(int x, int y, int64 time);
+  void PointSeen(float x, float y, int64 time);
   float XVelocity();
   float YVelocity();
   float VelocitySquared();
@@ -26,8 +26,8 @@ class EVENTS_EXPORT VelocityCalculator {
 
  private:
   struct Point {
-    int x;
-    int y;
+    float x;
+    float y;
     int64 time;
   };
